// src/types/database.ts
export type ParentRole = 'madre' | 'padre' | 'tutor';
export type ParentingStyle = 'autoritario' | 'democratico' | 'permisivo' | 'negligente' | 'no_clasificado';